Added User module 3 task

+++ Discussion-App/script.js
@@ -32,7 +32,7 @@ function searchQuestion(){
         leftDataListNode.innerHTML = "";
         allQuestions.forEach((question)=>{
             if((question.title).includes(searchedElement)){
-                createQuestionList(question);
+                createQuestionList(question, 1);
                 flag = 1;
             }
         });
@@ -41,7 +41,7 @@ function searchQuestion(){
                 title : "No data found!",
                 description : ""
             }
-            createQuestionList(q);
+            createQuestionList(q, 2);
         }
 
     }
@@ -53,10 +53,10 @@ function searchQuestion(){
 
 //main function
 function init(){
-    let allSavedQuestions = savedQuestions();
+    let allSavedQuestions = sortTheQuestionList();
     if(allSavedQuestions != []){
         allSavedQuestions.forEach(function(question) {
-            createQuestionList(question);
+            createQuestionList(question, 1);
         });
     }
 }
@@ -73,12 +73,14 @@ function onSaveQuestion(){
         description : rightQuestionNode.value,
         responses : [],
         upvote : 0,
-        downvote : 0
+        downvote : 0,
+        timestamp : (new Date()),
+        favorite : false,
     };
     removeInput();
     if(question.title != "" && question.description != "")
     {
-        createQuestionList(question);
+        createQuestionList(question, 1);
         saveIntoLocalStorage(question);
     }
 }
@@ -95,20 +97,38 @@ function createId(){
 }
 
 //create the container for question on left side
-function createQuestionList(question){
+function createQuestionList(question, flag){
     var questionContainer = document.createElement("div");
     questionContainer.style.backgroundColor = "white";
     var questionTitle = document.createElement("h3");
     questionContainer.setAttribute("id", `${question.questionId}`);
     var questionDescription = document.createElement("p");
+    if(flag == 1)
+    {
+
+        var timestampNode = document.createElement("span");
+        var favoriteBtn = document.createElement("button");
+        favoriteBtn.innerHTML = "Favorite";
+        favoriteBtn.style.float = "right";
+        timestampNode.innerHTML = timeAgo(question.timestamp);
+        timestampNode.style.fontSize = "9px";
+        timestampNode.style.float = "right";
+        questionContainer.appendChild(favoriteBtn);
+        favoriteBtn.onclick = addToFavorite(question);
+    }
     questionTitle.innerHTML = question.title;
     questionDescription.innerHTML = question.description;
     questionContainer.appendChild(questionTitle);
+    if(flag == 1)
+    {
+        questionContainer.appendChild(timestampNode);
+    }
     questionContainer.appendChild(questionDescription);
 
     leftDataListNode.appendChild(questionContainer);
 
     questionContainer.addEventListener("click", onQuestionClick(question));
+
     
 }
 
@@ -169,6 +189,7 @@ function showQuestionOnRightPanel(question){
 }
 
 var responseListNode = document.createElement("div");
+
 //check for the comments
 function showResponses(question){
     var responsesTitle =  document.createElement("h2");
@@ -176,6 +197,7 @@ function showResponses(question){
     rightResponseAnswerNode.appendChild(responsesTitle);
     if((question.responses).length != 0){
         question.responses.forEach(function(response){
+            console.log(response);
             createResponsesList(response);
         })
     }
@@ -191,9 +213,14 @@ function createResponsesList(response){
     var commentNodeList = document.createElement("div");
     var nameNode = document.createElement("h3");
     var commentNode = document.createElement("p");
+    var timestampNode = document.createElement("span");
+    timestampNode.innerHTML = timeAgo(response.timestamp);
+    timestampNode.style.fontSize = "9px";
+    timestampNode.style.float = "right";
     nameNode.innerHTML = response.title;
     commentNode.innerHTML = response.comment;
     commentNodeList.appendChild(nameNode);
+    commentNodeList.appendChild(timestampNode);
     commentNodeList.appendChild(commentNode);
     rightResponseAnswerNode.appendChild(commentNodeList);
 }
@@ -222,22 +249,25 @@ function saveResponses(question){
     return function(){
         let response = {
             title : commentatorName.value,
-            comment : commentatorComment.value
+            comment : commentatorComment.value,
+            timestamp : (new Date()),
         }
-        createResponsesList(response);
-        question.responses.push(response);
-        let allQuestions = savedQuestions();
-        const key = question.questionId;
-        console.log(question,key);
-        let allNonRepondedQuestion = allQuestions.filter(element => {
-            if(element.questionId != key){
-                return true;
-            }
-        });
-        allNonRepondedQuestion.push(question);
-        localStorage.clear();
-        localStorage.setItem("questions", JSON.stringify(allNonRepondedQuestion));
-        removeResponseInputs();
+        if(response.title.length > 0 && response.comment.length > 0){
+            createResponsesList(response);
+            question.responses.push(response);
+            let allQuestions = savedQuestions();
+            const key = question.questionId;
+            console.log(question,key);
+            let allNonRepondedQuestion = allQuestions.filter(element => {
+                if(element.questionId != key){
+                    return true;
+                }
+            });
+            allNonRepondedQuestion.push(question);
+            localStorage.clear();
+            localStorage.setItem("questions", JSON.stringify(allNonRepondedQuestion));
+            removeResponseInputs();
+        }
     }
 }
 
@@ -253,13 +283,13 @@ function onUpvoteClick(question){
         if( allQuestions.length ){
             var finalQuestions = allQuestions.filter((ques)=>{
                 if(ques.questionId != question.questionId){
-                    return true
+                    return true;
                 }
             });
             localStorage.clear();
             var upvotedQuestion = allQuestions.filter((ques)=>{
                 if(ques.questionId == question.questionId){
-                    return true
+                    return true;
                 }
             });
             upvotedQuestion[0].upvote = upvotedQuestion[0].upvote + 1;
@@ -277,13 +307,13 @@ function onDownvoteClick(question){
         if( allQuestions.length ){
             var finalQuestions = allQuestions.filter((ques)=>{
                 if(ques.questionId != question.questionId){
-                    return true
+                    return true;
                 }
             });
             localStorage.clear();
             var downvotedQuestion = allQuestions.filter((ques)=>{
                 if(ques.questionId == question.questionId){
-                    return true
+                    return true;
                 }
             });
             downvotedQuestion[0].downvote = downvotedQuestion[0].downvote + 1;
@@ -341,3 +371,116 @@ function showDownvoteValue(downvote){
 newQuestionFormButton.addEventListener("click", function(){
     hideRightQuestion();
 })
+
+function sortTheQuestionList(){
+    let allQuestions = savedQuestions();
+    allQuestions.sort(function(a,b){
+        return b.upvote - a.upvote;
+    })
+    return allQuestions;
+}
+
+
+
+const MONTH_NAMES = [
+  'January', 'February', 'March', 'April', 'May', 'June',
+  'July', 'August', 'September', 'October', 'November', 'December'
+];
+
+
+function getFormattedDate(date, prefomattedDate = false, hideYear = false) {
+  const day = date.getDate();
+  const month = MONTH_NAMES[date.getMonth()];
+  const year = date.getFullYear();
+  const hours = date.getHours();
+  let minutes = date.getMinutes();
+
+  if (minutes < 10) {
+    // Adding leading zero to minutes
+    minutes = `0${ minutes }`;
+  }
+
+  if (prefomattedDate) {
+    // Today at 10:20
+    // Yesterday at 10:20
+    return `${ prefomattedDate } at ${ hours }:${ minutes }`;
+  }
+
+  if (hideYear) {
+    // 10. January at 10:20
+    return `${ day }. ${ month } at ${ hours }:${ minutes }`;
+  }
+
+  // 10. January 2017. at 10:20
+  return `${ day }. ${ month } ${ year }. at ${ hours }:${ minutes }`;
+}
+
+
+// --- Main function
+function timeAgo(dateParam) {
+  if (!dateParam) {
+    return null;
+  }
+
+  const date = typeof dateParam === 'object' ? dateParam : new Date(dateParam);
+  const DAY_IN_MS = 86400000; // 24 * 60 * 60 * 1000
+  const today = new Date();
+  const yesterday = new Date(today - DAY_IN_MS);
+  const seconds = Math.round((today - date) / 1000);
+  const minutes = Math.round(seconds / 60);
+  const isToday = today.toDateString() === date.toDateString();
+  const isYesterday = yesterday.toDateString() === date.toDateString();
+  const isThisYear = today.getFullYear() === date.getFullYear();
+
+
+  if (seconds < 5) {
+    return 'now';
+  } else if (seconds < 60) {
+    return `${ seconds } seconds ago`;
+  } else if (seconds < 90) {
+    return 'about a minute ago';
+  } else if (minutes < 60) {
+    return `${ minutes } minutes ago`;
+  } else if (isToday) {
+    return getFormattedDate(date, 'Today'); // Today at 10:20
+  } else if (isYesterday) {
+    return getFormattedDate(date, 'Yesterday'); // Yesterday at 10:20
+  } else if (isThisYear) {
+    return getFormattedDate(date, false, true); // 10. January at 10:20
+  }
+
+  return getFormattedDate(date); // 10. January 2017. at 10:20
+}
+
+function addToFavorite(question){
+    return ()=>{
+        console.log(question);
+        if(question.favorite){
+            question.favorite = false;
+        }
+        else{
+            question.favorite = true;
+        }
+        changeFavoriteColor(question);
+        localStorage.clear();
+        let allQuestions = savedQuestions();
+        let remQuestions = allQuestions.filter(function(que){
+            if(que.questionId != question.questionId){
+                return true
+            }
+        });
+        console.log(remQuestions);
+        remQuestions.push(question);
+        localStorage.setItem("questions",JSON.stringify(remQuestions));
+    }
+}
+
+function changeFavoriteColor(question){
+    var favoriteBtn = document.getElementById(`${question.questionId}`).children[0];
+    if(question.favorite){
+        favoriteBtn.style.backgroundColor = "blue";
+    }
+    else{
+        favoriteBtn.style.backgroundColor = "white";
+    }
+}
